Lab6 - DeiT-S Pipeline Parallelism/requirements.txt
torch==2.3.0
torchvision==0.18.0
numpy==1.24.0
tqdm==4.66.1
timm==0.9.16
torchpippy==0.2.0